# api/app/openapi/OpenAPI_backend_En.yaml
openapi: 3.1.0
info:
  title: Dev Portal & Translation → Chroma Pipeline API
  version: 0.3.0
  description: >
    An API that provides the pipeline: Extract → Translate → (optional Writeback) → Packaging → Chroma upsert,
    plus CRUD/import operations for the portal_* tables.
    Reflects Phase J/K policies and clarifies POST idempotency / Request-ID / 429/409 responses.
...
servers:
  - url: https://api.example.com
    description: default (switch prod/dev via env overlay)
security: []  # In the prod overlay, apply security: [{ bearerAuth: [] }]
...
paths:
...
  /portal/model:
    get:
      tags: [Portal Model]
      summary: List models
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalModelList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal Model]
      summary: Create model
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalModelCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalModel' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
...
  /portal/field:
    get:
      tags: [Portal Field]
      summary: List fields
      parameters:
        - in: query
          name: model
          schema: { type: string }
        - in: query
          name: field_name
          schema: { type: string }
        - in: query
          name: origin
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalFieldList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal Field]
      summary: Create field
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalFieldCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalField' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
...
  /portal/field/extract:
    post:
      x-dev-only: true
      tags: [Portal Field, Extract]
      summary: Enqueue specified fields to translate (UPSERT)
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ExtractFieldRequest' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/ExtractResult' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }

  ############################################
  # Portal: view_common (action-centric)
  ############################################
  /portal/view_common:
    get:
      tags: [Portal View Common]
      summary: List
      parameters:
        - in: query
          name: action_xmlid
          schema: { type: string }
        - in: query
          name: model
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalViewCommonList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal View Common]
      summary: Create
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action_xmlid: { type: string }
                model: { type: string }
                action_name: { type: string }
                model_label: { type: string }
                model_table: { type: string }
                view_types: { type: array, items: { type: string } }
                primary_view_type: { type: string }
                help_ja_text: { type: string }
                help_en_text: { type: string }
                ai_purpose: { type: string }
                ai_purpose_i18n: { type: object, additionalProperties: true }
              required: [action_xmlid, model]
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalViewCommon' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
...
  /portal/view_common/extract:
    post:
      x-dev-only: true
      tags: [Portal View Common, Extract]
      summary: Enqueue ai_purpose / help_ja_text to translate (only where JA is present & EN is empty)
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ExtractViewCommonRequest' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/ExtractResult' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }

  ############################################
  # Portal: view (type-specific)
  ############################################
  /portal/view:
    get:
      tags: [Portal View]
      summary: List
      parameters:
        - in: query
          name: common_id
          schema: { type: integer }
        - in: query
          name: view_type
          schema: { type: string, enum: [form, kanban, list, calendar, search, graph, pivot, dashboard, tree, map] }
        - in: query
          name: model
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalViewList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal View]
      summary: Create
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalViewCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalView' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
...
  /portal/tab:
    get:
      tags: [Portal Tab]
      summary: List tabs
      parameters:
        - in: query
          name: view_id
          schema: { type: integer }
        - in: query
          name: tab_key
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalTabList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal Tab]
      summary: Create
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalTabCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalTab' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
...
  /portal/smart_button:
    get:
      tags: [Portal Smart Button]
      summary: List smart buttons
      parameters:
        - in: query
          name: view_id
          schema: { type: integer }
        - in: query
          name: button_key
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalSmartButtonList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal Smart Button]
      summary: Create
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalSmartButtonCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalSmartButton' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
...
  /portal/menu:
    get:
      tags: [Portal Menu]
      summary: List menus
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalMenuList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal Menu]
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalMenuCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalMenu' } } }
...
  /translate:
    get:
      x-dev-only: true
      tags: [Translate]
      summary: List translate rows (dev)
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, translated, ready_for_chroma, done, failed] }
        - in: query
          name: entity
          schema: { $ref: '#/components/schemas/Entity' }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/TranslateList' } } }
...
  /chroma/docs:
    get:
      x-dev-only: true
      tags: [Package]
      summary: List portal_chroma_doc (dev)
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [queued, upserted, failed] }
        - in: query
          name: entity
          schema: { $ref: '#/components/schemas/Entity' }
        - in: query
          name: model
          schema: { type: string }
        - in: query
          name: collection
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
        - in: query
          name: cursor
          schema:
            type: string
            description: "Keyset cursor (base-10 string of id). Pass back the response's next_cursor as-is."
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/ChromaDocsList' } } }

  ############################################
  # Chroma upsert (Cron internal)
  ############################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Correlation ID per run. Generated by Cron/clients and attached to every request.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 256 }
      description: >
        Idempotency key for POST/PUT. Deterministically generated from a time window and payload.
        Matching retries return 200/409 (implementation-dependent) with no side effects.
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ConflictIdempotent:
      description: Treated as replay due to idempotency
      headers:
        Idempotency-Status:
          description: If 'replayed', the request was treated as a replay
          schema: { type: string, enum: [replayed, conflict] }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        detail: { type: string }
        status: { type: integer }
      required: [title, status]
    NaturalKey:
      type: string
      description: "Convention: entity::... e.g., field::sale.order::partner_id"
    Entity:
      type: string
      enum: [field, view_common, model, tab, smart_button, menu]
    Lang:
      type: string
      enum: [ja, en]
    UpsertMode:
      type: string
      enum: [upsert, skip_existing, upsert_if_changed]
      default: upsert_if_changed
...
    CursorList:
      type: object
      properties:
        items: { type: array, items: {} }
        next_cursor: { type: string, nullable: true }

    ########################################
    # Portal model
    ########################################
    PortalModel:
      type: object
      properties:
        id: { type: integer }
        model: { type: string }
        model_table: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalModelCreate:
      type: object
      properties:
        model: { type: string }
        model_table: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        notes: { type: string }
      required: [model, model_table]
...
    PortalModelList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalModel' } }

    ########################################
    # Portal field
    ########################################
    PortalField:
      type: object
      properties:
        id: { type: integer }
        model_id: { type: integer }
        model: { type: string }
        model_table: { type: string }
        field_name: { type: string }
        ttype: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        notes: { type: string, nullable: true }
        origin: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalFieldCreate:
      type: object
      properties:
        model: { type: string }
        field_name: { type: string }
        ttype: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        notes: { type: string }
      required: [model, field_name, ttype]
...
    PortalFieldList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalField' } }

    ########################################
    # Portal view_common
    ########################################
    PortalViewCommon:
      type: object
      properties:
        id: { type: integer }
        action_xmlid: { type: string }
        action_name: { type: string, nullable: true }
        model: { type: string }
        model_label: { type: string, nullable: true }
        model_table: { type: string }
        view_types: { type: array, items: { type: string } }
        primary_view_type: { type: string }
        help_ja_text: { type: string, nullable: true }
        help_en_text: { type: string, nullable: true }
        ai_purpose: { type: string, nullable: true }
        ai_purpose_i18n: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    PortalViewCommonList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalViewCommon' } }

    ########################################
    # Portal view
    ########################################
    PortalView:
      type: object
      properties:
        id: { type: integer }
        common_id: { type: integer }
        view_type: { type: string }
        model: { type: string }
        view_name: { type: string, nullable: true }
        is_primary: { type: boolean }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalViewCreate:
      type: object
      properties:
        common_id: { type: integer }
        view_type: { type: string }
        model: { type: string }
      required: [common_id, view_type]
...
    PortalViewList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalView' } }

    ########################################
    # Portal tab
    ########################################
    PortalTab:
      type: object
      properties:
        id: { type: integer }
        view_id: { type: integer }
        tab_key: { type: string }
        tab_label_ja: { type: string, nullable: true }
        tab_label_en: { type: string, nullable: true }
        model: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalTabCreate:
      type: object
      properties:
        view_id: { type: integer }
        tab_key: { type: string }
        tab_label_ja: { type: string }
        tab_label_en: { type: string }
        model: { type: string }
      required: [view_id, tab_key]
...
    PortalTabList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalTab' } }
...
    PortalSmartButton:
      type: object
      properties:
        id: { type: integer }
        view_id: { type: integer }
        button_key: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        model: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalSmartButtonCreate:
      type: object
      properties:
        view_id: { type: integer }
        button_key: { type: string }
        label_i18n: { type: object, additionalProperties: true }
        model: { type: string }
      required: [view_id, button_key]
...
    PortalSmartButtonList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalSmartButton' } }
...
    PortalMenu:
      type: object
      properties:
        id: { type: integer }
        menu_xmlid: { type: string, nullable: true }
        name_ja: { type: string, nullable: true }
        name_en: { type: string, nullable: true }
        name_i18n: { type: object, additionalProperties: true }
        model: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalMenuCreate:
      type: object
      properties:
        menu_xmlid: { type: string }
        name_i18n: { type: object, additionalProperties: true }
        model: { type: string }
...
    PortalMenuList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalMenu' } }
...
    ExtractViewCommonRequest:
      type: object
      properties:
        action_xmlids: { type: array, items: { type: string } }
        targets:
          type: array
          items: { type: string, enum: [ai_purpose, help] }
        mode: { $ref: '#/components/schemas/UpsertMode' }
      required: [action_xmlids]
    ExtractFieldRequest:
      type: object
      properties:
        models: { type: array, items: { type: string } }
        fields: { type: array, items: { type: string } }
        mode: { $ref: '#/components/schemas/UpsertMode' }
    ExtractResult:
      type: object
      properties:
        picked: { type: integer }
        inserted: { type: integer }
        updated: { type: integer }
        skipped_no_ja: { type: integer }
        skipped_has_en: { type: integer }
        skipped_not_found: { type: integer }
        details:
          type: array
          items:
            type: object
            properties:
              natural_key: { $ref: '#/components/schemas/NaturalKey' }
              reason: { type: string }

    ########################################
    # Translate
    ########################################
...
    TranslateRow:
      type: object
      properties:
        natural_key: { $ref: '#/components/schemas/NaturalKey' }
        entity: { $ref: '#/components/schemas/Entity' }
        model: { type: string }
        label: { type: string }
        purpose: { type: string }
        translated_label: { type: string }
        translated_purpose: { type: string }
        status: { type: string, enum: [pending, translated, ready_for_chroma, done, failed] }
        updated_at: { type: string, format: date-time }
    TranslateList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/TranslateRow' } }

    ########################################
    # Writeback
    ########################################
...
    ChromaDoc:
      type: object
      description: "doc_text is capped at 16 KiB safe length (UTF-8 safe truncation)"
      properties:
        doc_id: { type: string }
        natural_key: { $ref: '#/components/schemas/NaturalKey' }
        lang: { $ref: '#/components/schemas/Lang' }
        collection: { type: string }
        doc_text: { type: string, maxLength: 16384 }
        entity: { $ref: '#/components/schemas/Entity' }
        model: { type: string }
        model_table: { type: string }
        field_name: { type: string, nullable: true }
        action_xmlid: { type: string, nullable: true }
        target: { type: string, nullable: true }
        metadata: { type: object, additionalProperties: true }
        status: { type: string, enum: [queued, upserted, failed] }
        updated_at: { type: string, format: date-time }
    ChromaDocsList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/ChromaDoc' } }

    ########################################
    # Chroma upsert
    ########################################
...